Outputs
Which is the maximum range: 10

List of primes in range [0, 10] is:
[2, 3, 5, 7]

The product of prime numbers in range [1, 10] is:
210

List of highest primes powers in range [0, 10] is:
[5, 7, 8, 9]

The smallest positive number that is evenly divisible by all of the numbers from 1 to 10 is:
2,520

Do you want to generate a new prime? Enter Yes or No: y
Which is the maximum range: 20

List of primes in range [0, 20] is:
[2, 3, 5, 7, 11, 13, 17, 19]

The product of prime numbers in range [1, 20] is:
9,699,690

List of highest primes powers in range [0, 20] is:
[5, 7, 9, 11, 13, 16, 17, 19]

The smallest positive number that is evenly divisible by all of the numbers from 1 to 20 is:
232,792,560

Do you want to generate a new prime? Enter Yes or No: y
Which is the maximum range: 30

List of primes in range [0, 30] is:
[2, 3, 5, 7, 11, 13, 17, 19, 23, 29]

The product of prime numbers in range [1, 30] is:
6,469,693,230

List of highest primes powers in range [0, 30] is:
[7, 11, 13, 16, 17, 19, 23, 25, 27, 29]

The smallest positive number that is evenly divisible by all of the numbers from 1 to 30 is:
2,329,089,562,800

Do you want to generate a new prime? Enter Yes or No: y
Which is the maximum range: 40

List of primes in range [0, 40] is:
[2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37]

The product of prime numbers in range [1, 40] is:
7,420,738,134,810

List of highest primes powers in range [0, 40] is:
[7, 11, 13, 17, 19, 23, 25, 27, 29, 31, 32, 37]

The smallest positive number that is evenly divisible by all of the numbers from 1 to 40 is:
5,342,931,457,063,200

Do you want to generate a new prime? Enter Yes or No: n